style(lint): apply ruff autofix + format across the repo (415 files)
CI's Lint & Type Check job has never once reached the ruff steps in its
history -- pip install always failed first (fixed in 7e4d14f). Unblocking
that revealed this codebase has never had ruff enforced against it:
1226 lint errors and 393 of ~568 files not matching `ruff format`,
entirely pre-existing, unrelated to today's other work.

Applied in three passes, each re-verified before the next:
  1. `ruff check --fix` (safe fixes): 514 resolved -- import sorting,
     Optional[X] -> X | None, datetime.timezone.utc -> datetime.UTC,
     unnecessary noqa removal, and similar zero-semantic-risk rewrites.
  2. `ruff check --fix --unsafe-fixes`: 539 more resolved -- the highest-
     risk category here is UP042 (class X(str, Enum) -> class X(StrEnum),
     8 classes: JobStatus, BrokerMode, OrderStatus, OrderType,
     ValidationStatus, OperationalStatus, EconomicStatus, SampleAdequacy).
     This changes str()/f-string output (confirmed empirically: old style
     stringifies as "ClassName.MEMBER", StrEnum as the bare value) --
     directly relevant here because 4 of those 8 enums feed
     research/forward_validation/models.py's blake2b fingerprint payload,
     and this codebase's fingerprints are relied on as reproducibility
     guarantees throughout (M7/M22/M23/M24). Traced the actual
     serialization path (_fingerprint_payload -> _fp -> _canonical ->
     json.dumps(..., default=str)) and confirmed empirically that
     json.dumps encodes str-subclass instances (which both the old and
     new enum styles are) by their raw character data, never invoking
     __str__ -- verified byte-identical output for both styles before
     trusting this. The full suite passing unchanged (below) is
     consistent with, not a substitute for, that mechanistic check.
  3. `ruff format`: 380 files reformatted, 309 unchanged. Whitespace/line-
     wrapping only.

239 errors remain -- everything ruff itself cannot safely auto-fix
(N806/N803/N812/N814/N805 naming, PT009/PT001/PT012/PT006 pytest-style,
B905/B904/B023/B017/B009/B007 correctness patterns needing a real code
read, RUF012 mutable class defaults, E702/E741/E731/E402 style, and 13
more categories). `ruff check` still exits non-zero, so CI's Lint job
will still fail at that step until those are fixed by hand -- this
commit does not turn CI green by itself, only reduces 1226 -> 239 and
resolves formatting entirely.

Verified: full local suite unchanged before and after every pass --
2924 passed, 3 skipped, 6 deselected, 2 warnings, 0 failed.

No functional files touched beyond what ruff rewrote; no manual edits
in this commit.

+++ src/mentisrex/research/validation/hac.py
@@ -23,7 +23,7 @@ def auto_lag(n: int) -> int:
     """Newey-West (1994) automatic Bartlett lag. Non-negative, < n."""
     if n < 2:
         return 0
-    return min(n - 1, int(math.floor(4.0 * (n / 100.0) ** (2.0 / 9.0))))
+    return min(n - 1, math.floor(4.0 * (n / 100.0) ** (2.0 / 9.0)))
 
 
 def hac_long_run_variance(returns, lag: int | None = None) -> float:
@@ -72,9 +72,16 @@ def hac_significance(returns, lag: int | None = None) -> dict:
     used_lag = auto_lag(n) if lag is None else max(0, min(lag, max(n - 1, 0)))
     se = hac_standard_error(r, lag)
     if n < 2 or not math.isfinite(se) or se == 0.0:
-        return {"hac_se": 0.0 if n >= 2 else float("nan"),
-                "hac_t_stat": 0.0, "hac_p_value": 1.0, "hac_lag": used_lag}
+        return {
+            "hac_se": 0.0 if n >= 2 else float("nan"),
+            "hac_t_stat": 0.0,
+            "hac_p_value": 1.0,
+            "hac_lag": used_lag,
+        }
     t = float(r.mean()) / se
-    return {"hac_se": float(se), "hac_t_stat": float(t),
-            "hac_p_value": float(student_t_two_sided_p(t, n - 1)),
-            "hac_lag": used_lag}
+    return {
+        "hac_se": float(se),
+        "hac_t_stat": float(t),
+        "hac_p_value": float(student_t_two_sided_p(t, n - 1)),
+        "hac_lag": used_lag,
+    }
